[feat] reply tag

+++ controllers/replycomments.controller.js
@@ -10,7 +10,7 @@ const reCommentSchema = Joi.object({
 class ReplyCommentController {
   replycommentService = new ReplyCommentService();
 
-   //대댓글 작성
+   //답글 작성
   createReComment = async (req, res, next) => {
     try {
       const { postIdx, commentIdx } = req.params;
@@ -19,7 +19,7 @@ class ReplyCommentController {
 
       const resultSchema = await reCommentSchema.validate({ comment });
       if (resultSchema.error && comment.length < 1 ) {
-        throw Boom.preconditionFailed('대댓글 내용을 입력해주세요.'); //412
+        throw Boom.preconditionFailed('답글 내용을 입력해주세요.'); //412
       }
 
       const seletPost = await this.replycommentService.findPost(postIdx)
@@ -33,22 +33,21 @@ class ReplyCommentController {
       }
 
       await this.replycommentService.createReComment(
-        postIdx,
         commentIdx,
+        postIdx,
         userIdx,
         comment
       );
 
       return res.status(200).json({
-          message: '대댓글을 작성하였습니다.',
+          message: '답글을 작성하였습니다.',
       });
     } catch (error) {
-        logger.error(error.message);
-        throw error;
-      }
+      next(error)
+    }
   };
 
-   //대댓글 조회
+   //답글 조회
   getReComments = async (req, res, next) => {
     try {
       const { postIdx, commentIdx } = req.params;
@@ -69,13 +68,12 @@ class ReplyCommentController {
         replys: seletReComments,
       });
     } catch (error) {
-        logger.error(error.message);
-        throw error;
-      }
+      next(error)
+    }
   };
 
 
-   //대댓글 수정
+   //답글 수정
   updateReComment = async (req, res, next) => {
     try {
       const { commentIdx, postIdx, replyIdx } = req.params;
@@ -94,31 +92,30 @@ class ReplyCommentController {
 
       const resultSchema = await reCommentSchema.validate({ comment });
       if (resultSchema.error && comment.length < 1 ) {
-        throw Boom.preconditionFailed('대댓글 내용을 입력해주세요.');
+        throw Boom.preconditionFailed('답글 내용을 입력해주세요.');
       }
 
-      const chkAuthComment = await this.replycommentService.findeAuth( replyIdx, userIdx );
+      const chkAuthComment = await this.replycommentService.findAuth( replyIdx, userIdx );
       if (!chkAuthComment) {
-        throw Boom.forbidden('대댓글의 수정 권한이 없습니다.'); //403
+        throw Boom.forbidden('답글의 수정 권한이 없습니다.'); 
       }
 
-      await this.commentService.updateComment(
+      await this.replycommentService.updateReComment(
         replyIdx,
         comment,
         userIdx
       );
 
       return res.status(200).json({
-        message: '대댓글을 수정하였습니다.',
+        message: '답글을 수정하였습니다.',
       });
-    } catch (err) {
-        logger.error(error.message);
-        throw error;
+    } catch (error) {
+      next(error)
     }
   };
 
 
-   //대댓글 삭제
+   //답글 삭제
   deleteReComment = async (req, res, next) => {
     try {
       const { commentIdx, postIdx, replyIdx } = req.params;
@@ -129,23 +126,26 @@ class ReplyCommentController {
         throw Boom.notFound('해당 게시글은 존재하지 않습니다.');
       }
 
-      const chkAuthComment = await this.replycommentService.findeAuth( replyIdx, userIdx );
+      const seletComment = await this.replycommentService.findComment(commentIdx)
+      if (!seletComment) {
+        throw Boom.notFound('해당 댓글은 존재하지 않습니다.'); 
+      }
+
+      const chkAuthComment = await this.replycommentService.findAuth(replyIdx, userIdx);
       if (!chkAuthComment) {
-        throw Boom.forbidden('댓글의 삭제 권한이 없습니다.');
+        throw Boom.forbidden('답글의 삭제 권한이 없습니다.');
       }
 
       await this.replycommentService.deleteReComment(
-        replyIdx,
-        userIdx
+        replyIdx
       );
 
       return res.status(200).json({
-        message: '댓글을 삭제하였습니다.',
+        message: '답글을 삭제하였습니다.',
       });
-    } catch (err) {
-      logger.error(error.message);
-      throw error;
-  }
+    } catch (error) {
+      next(error)
+    }
   };
 }
